Separate connection, devteam, and contacts
