+++ api/connection/index.js
@@ -1,0 +1,24 @@
+const mariadb = require('mariadb');
+
+require('dotenv-extended').load({
+    encoding: 'utf8',
+    silent: true,
+    path: '.env',
+    defaults: '.env.defaults',
+    schema: '.env.schema',
+    errorOnMissing: false,
+    errorOnExtra: false,
+    assignToProcessEnv: true,
+    overrideProcessEnv: false
+  });
+
+// Connection stuff
+const connection = mariadb.createPool({
+    host: 'localhost',
+    user: process.env.DB_USER,
+    password: process.env.DB_PASSWORD,
+    database: 'connecc',
+    connectionLimit: 5,
+});
+
+module.exports = connection
