+++ index.js
@@ -1,10 +1,9 @@
-const express = require(`express`);
-const bodyParser = require(`body-parser`);
-const cors = require(`cors`);
+const express = require('express');
+const bodyParser = require('body-parser');
+const cors = require('cors');
 
-const extModule = require(`./api/index.js`);
-const devTeam = extModule.devTeam
-const contact = extModule.contacts
+const devTeam = require('./api/devteam')
+const contact = require('./api/contacts')
 
 const server = express();
 const port = 3000;
@@ -22,8 +21,9 @@ server.get(`/contacts/search`, contact.searchContact);
 server.post(`/contacts`, contact.addContact);
 server.delete(`/contacts/:id`, contact.deleteContact);
 server.get(`/devteam`, devTeam.showDevTeam);
+
 server.listen(port, () =>
   console.log(`
-        Contact server listening on port ${port}
+        Server listening on port ${port}
         `)
 );
